clk: Ingenic: Improve CPU clock related code.
For the JZ4780 SoC, the X1000 SoC, and the X1830 SoC, the clock frequencies of
their CPU and L2 Cache must meet a specific ratio, so when adjusting the clock
frequency divider factor of CPU, the clock frequency divider factor of L2 Cache
needs to be adjusted accordingly. The current driver does not take this detail
into account, so there is a high probability that after adjusting the CPU clock
frequency divider factor, the ratio between the clock frequency of CPU and the
clock frequency of L2 Cache does not meet the requirement and leads to a crash.
With this patch, this issue will be resolved.

Signed-off-by: 周琰杰 (Zhou Yanjie) <zhouyanjie@wanyeetech.com>